requirements.txt
asgiref==3.3.4
attrs==21.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
distro==1.5.0
Django==3.2.2
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-pytest==0.2.0
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.12.4
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
easy-thumbnails==2.7.1
easy-thumbnails-rest==1.1.2
Faker==8.1.4
flake8==3.9.2
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
jsonschema==3.2.0
mccabe==0.6.1
packaging==20.9
paramiko==2.7.2
Pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-django==4.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
websocket-client==0.59.0
xlwt==1.3.0
zipp==3.4.1
